application/requirements.txt
customtkinter
aiohttp
Pillow
requests
python-socketio
opencv-python
numpy
mss
av
aiortc
sounddevice
cryptography
python-dotenv
pystray